Fixes to analyzers REGEX
