--- Pobie.Roslyn.Analyzers/AnalyzerReleases.Shipped.md
@@ -1,6 +1,0 @@
-## Release 1.0
-
-### New Rules
-
-Rule ID | Category | Severity | Notes
---------|----------|----------|--------------------
