+++ Pobie.Roslyn.Tests/Analyzers/FluentAssertions/BecauseStartsWithKeywordAnalyzerTest.cs
@@ -1,25 +1,16 @@
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Pobie.Analyzers.Analyzers.FluentAssertions;
+using Pobie.Roslyn.Analyzers.FluentAssertions;
 using Xunit;
 
-namespace Analyzers.Tests.Analyzers.FluentAssertions;
+namespace Pobie.Roslyn.Tests.Analyzers.FluentAssertions;
 
-public class BecauseDoesNotEndWithPunctuationAnalyzerTest
-    : AnalyzerTest<BecauseDoesNotEndWithPunctuationAnalyzer>
+public class BecauseStartsWithKeywordAnalyzerTest : AnalyzerTest<BecauseStartsWithKeywordAnalyzer>
 {
-    [Theory]
-    [InlineData(".")]
-    [InlineData(",")]
-    [InlineData("!")]
-    [InlineData("?")]
-    [InlineData(";")]
-    [InlineData(":")]
-    public async Task ShouldWarn_WhenCallingBooleanAssertion_GivenPunctuationAtTheEndOfBecause(
-        string punctuation
-    )
+    [Fact]
+    public async Task ShouldWarn_WhenCallingBooleanAssertion_GivenBecauseDoesNotStartWithKeyword()
     {
-        string assertionWithPunctuation = $$"""
+        const string assertionWithPunctuation = """
             using FluentAssertions;
             using Xunit;
 
@@ -29,13 +20,13 @@ public class BecauseDoesNotEndWithPunctuationAnalyzerTest
                 public void ShouldReturnTrue()
                 {
                     var aValue = true;
-                    aValue.Should().Be(true, "because it should be true{{punctuation}}");
+                    aValue.Should().Be(true, "it should be true.");
                 }
             }
             """;
 
         await ShouldHaveDiagnostic(
-            BecauseDoesNotEndWithPunctuationAnalyzer.Id,
+            BecauseStartsWithKeywordAnalyzer.Id,
             10,
             9,
             assertionWithPunctuation,
@@ -44,7 +35,7 @@ public class BecauseDoesNotEndWithPunctuationAnalyzerTest
     }
 
     [Fact]
-    public async Task ShouldNotWarn_WhenCallingBooleanAssertion_GivenNoPunctuationAtTheEndOfBecause()
+    public async Task ShouldNotWarn_WhenCallingBooleanAssertion_GivenBecauseStartsWithKeyword()
     {
         const string assertionWithoutPunctuation = """
             using FluentAssertions;
@@ -65,7 +56,7 @@ public class BecauseDoesNotEndWithPunctuationAnalyzerTest
     }
 
     [Fact]
-    public async Task ShouldNotWarn_WhenCallingBooleanAssertion_GivenEmptyString()
+    public async Task ShouldNotWarn_WhenCallingBooleanAssertion_GivenNoBecause()
     {
         const string assertionWithoutPunctuation = """
             using FluentAssertions;
@@ -77,7 +68,7 @@ public class BecauseDoesNotEndWithPunctuationAnalyzerTest
                 public void ShouldReturnTrue()
                 {
                     var aValue = true;
-                    aValue.Should().Be(true, "");
+                    aValue.Should().Be(true);
                 }
             }
             """;
