+++ Pobie.Roslyn.Tests/Pobie.Roslyn.Tests.csproj
@@ -2,9 +2,8 @@
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
         <Nullable>enable</Nullable>
+
         <IsPackable>false</IsPackable>
-        <RootNamespace>Analyzers.Tests</RootNamespace>
-        <OutputType>Library</OutputType>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="FluentAssertions" Version="8.2.0"/>
@@ -13,9 +12,15 @@
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.13.0"/>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0"/>
         <PackageReference Include="xunit" Version="2.4.2"/>
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5"/>
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+
     </ItemGroup>
+
     <ItemGroup>
-        <ProjectReference Include="..\Pobie.Roslyn.Analyzers\Pobie.Roslyn.Analyzers.csproj"/>
+        <ProjectReference Include="..\Pobie.Roslyn\Pobie.Roslyn.csproj"/>
     </ItemGroup>
+
 </Project>
