+++ Pobie.Roslyn/AnalyzerReleases.Shipped.md
@@ -1,0 +1,9 @@
+## Release 1.0
+
+### New Rules
+
+ Rule ID | Category | Severity | Notes                                           
+---------|----------|----------|-------------------------------------------------
+ UT0001  | Usage    | Warning  | UT0001_BecausePresenceAnalyzer                  
+ UT0002  | Usage    | Warning  | UT0002_BecauseStartsWithKeywordAnalyzer         
+ UT0003  | Usage    | Warning  | UT0003_BecauseDoesNotEndWithPunctuationAnalyzer
