+++ Pobie.Roslyn/Analyzers/FluentAssertions/BecauseStartsWithKeywordAnalyzer.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections.Immutable;
-using Analyzers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Pobie.Analyzers.Analyzers.Parameter;
-using Pobie.Analyzers.Validations;
+using Pobie.Roslyn.Analyzers.Parameter;
+using Pobie.Roslyn.Validations;
 
-namespace Pobie.Analyzers.Analyzers.FluentAssertions;
+namespace Pobie.Roslyn.Analyzers.FluentAssertions;
 
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
 public class BecauseStartsWithKeywordAnalyzer : ParameterValueAnalyzer
@@ -26,9 +25,11 @@ public class BecauseStartsWithKeywordAnalyzer : ParameterValueAnalyzer
     public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics =>
         ImmutableArray.Create(Rule);
 
-    protected override string TargetClassName => "*Assertions";
-    protected override string TargetMethodName => "*";
+    protected override string TargetClassName => ".*Assertions";
+    protected override string TargetMethodName => ".*";
     protected override string TargetParameterName => "because";
+
+    protected override string TargetModule => "FluentAssertions.dll";
 
     protected override DiagnosticDescriptor Rule => DiagnosticDescriptor;
     protected override IValidation Validation { get; } = new StringStartsWithValidation("because ");
