+++ Pobie.Roslyn/Pobie.Roslyn.csproj
@@ -6,8 +6,10 @@
         <LangVersion>latest</LangVersion>
         <EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
         <IsRoslynComponent>true</IsRoslynComponent>
-        <RootNamespace>Pobie.Analyzers</RootNamespace>
-        <PackageId>Pobie.Roslyn.Analyzers</PackageId>
+        <AssemblyName>Pobie.Roslyn</AssemblyName>
+        <RootNamespace>Pobie.Roslyn</RootNamespace>
+        <PackageId>Pobie.Roslyn</PackageId>
+        <PackageReadmeFile>readme.md</PackageReadmeFile>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.11.0">
@@ -17,6 +19,7 @@
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.3.0"/>
         <PackageReference Update="NETStandard.Library" PrivateAssets="all"/>
     </ItemGroup>
+
     <ItemGroup>
         <EmbeddedResource Update="Resources.resx">
             <Generator>ResXFileCodeGenerator</Generator>
