+++ Pobie.Roslyn/tools/install.ps1
