+++ Pobie.Roslyn/tools/uninstall.ps1
